pyproject.toml
[tool.poetry]
name = "rdo_dailies"
version = "0.3.0"
description = "A discord bot that posts daily challenges"
authors = ["Anton <gmt@freemail.ru>"]
packages = [
    { include = "rdo_dailies" },
]
exclude = ["**/*.po"]

[tool.poetry.scripts]
rdo-dailies-announcer = 'rdo_dailies.__main__:start'

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
beautifulsoup4 = "^4.9.3"
aiofiles = "^0.6.0"
python-decouple = "^3.3"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
